+++ KestrelMock.Tests/TestStartup.cs
@@ -1,21 +1,13 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Net.Http.Headers;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-using KestrelMock.Services;
+﻿using KestrelMock.Services;
+using KestrelMock.Settings;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace KestrelMock
 {
-	public class TestStartup
+    public class TestStartup
 	{
 		private readonly IConfiguration configuration;
 		private readonly IWebHostEnvironment hostingEnvironment;
