+++ KestrelMock/Domain/PathMapping.cs
@@ -1,0 +1,11 @@
+﻿using KestrelMock.Settings;
+using System.Collections.Concurrent;
+
+namespace KestrelMock.Domain
+{
+    public sealed class PathMapping : ConcurrentDictionary<string, HttpMockSetting>
+    {
+
+    }
+
+}
