+++ KestrelMock/Domain/PathMatchesRegexMapping.cs
@@ -1,0 +1,12 @@
+﻿using KestrelMock.Settings;
+using System.Collections.Concurrent;
+using System.Text.RegularExpressions;
+
+namespace KestrelMock.Domain
+{
+    public sealed class PathMatchesRegexMapping : ConcurrentDictionary<Regex, HttpMockSetting>
+    {
+
+    }
+
+}
