+++ KestrelMock/Domain/PathStartsWithMapping.cs
@@ -1,0 +1,11 @@
+﻿using KestrelMock.Settings;
+using System.Collections.Concurrent;
+
+namespace KestrelMock.Domain
+{
+    public sealed class PathStartsWithMapping : ConcurrentDictionary<string, HttpMockSetting>
+    {
+
+    }
+
+}
