+++ KestrelMock/Settings/ContentType.cs
@@ -1,4 +1,4 @@
-﻿namespace KestrelMock
+﻿namespace KestrelMock.Settings
 {
     public enum ContentType
     {
