+++ KestrelMock/Settings/MockConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace KestrelMock
+namespace KestrelMock.Settings
 {
     public class MockConfiguration : List<HttpMockSetting>
 	{
