+++ KestrelMock/Settings/Request.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace KestrelMock
+namespace KestrelMock.Settings
 {
 	public class Request
 	{
